wip: cart feature

+++ lib/features/domain/usecases/shipping/get_freight_quote_usecase.dart
@@ -55,10 +55,10 @@ class Items {
 }
 
 class Dimensions {
-  final int kiloGrams;
-  final int length;
-  final int width;
-  final int height;
+  final double kiloGrams;
+  final double length;
+  final double width;
+  final double height;
   final bool addWoodenPacking;
 
   Dimensions({
